refactor: 관련된 클래스끼리 다시 패키징
