+++ src/main/java/lotto/model/lotto/Lottos.java
@@ -1,11 +1,9 @@
-package lotto.model;
+package lotto.model.lotto;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import lotto.model.ball.BonusBall;
-import lotto.model.ball.WinningBalls;
 import lotto.model.prize.MatchResult;
 
 public class Lottos {
